Preserve input dtype in per-instance FFT/bias transforms; guard subsample target
Ghosting, Spike, Motion and BiasField computed in float32 and only cast
back to the input dtype for applied elements, so per-instance gating
could mix dtypes and break torch.cat for non-float32 batches. Cast every
per-element output back to the input dtype (matching the Blur
convention). Added a dtype-preservation regression test.

Also reject non-positive subsample targets in _subsample_for_quantile
with a clear ValueError now that subsample_size is caller-exposed.

+++ src/torchio/transforms/intensity/ghosting.py
@@ -206,4 +206,4 @@ def _add_ghosting(
         torch.fft.ifftshift(corrupted, dim=(-3, -2, -1)),
         dim=(-3, -2, -1),
     ).real
-    return result
+    return result.to(data.dtype)
